pyproject.toml
[tool.poetry]
name = "notebook"
version = "0.1.0"
description = "ITMO University programming hometask"
authors = ["Ruslan Babayev <rus.babaef@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openpyxl = "^3.1.5"
pyqt5-sip = "^12.16.0"
pyqt5-tools = "^5.15.9.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
